Mathematics
Welcome to my Mathematics in Java repository! This repository contains well-documented and tested code that provides a solid foundation for understanding and utilizing various mathematical operations and functions in Java.

Mathematics is a fundamental subject that is used extensively in computer science and programming. This implementation includes a variety of useful methods, such as calculating prime numbers, finding the greatest common divisor, and performing various trigonometric functions. The code is well-documented and includes helpful comments that explain how each part of the implementation works.

Whether you are just starting to learn about mathematics in programming or you are an experienced programmer looking to incorporate these techniques into your projects, this repository is an excellent resource. Feel free to fork the code and customize it to your needs, or use it as a starting point for your own implementation.

Thank you for checking out my Mathematics in Java repository!