Searching
Welcome to my Searching in Java repository! This repository contains well-documented and tested code that provides a solid foundation for understanding and utilizing various searching algorithms in Java.

Searching is a common operation in computer science and programming and involves finding a specific element or value within a collection of data. This implementation includes a variety of useful searching algorithms, such as linear search, binary search, and interpolation search. The code is well-documented and includes helpful comments that explain how each part of the implementation works.

Whether you are just starting to learn about searching algorithms or you are an experienced programmer looking to incorporate these techniques into your projects, this repository is an excellent resource. Feel free to fork the code and customize it to your needs, or use it as a starting point for your own implementation.

Thank you for checking out my Searching in Java repository!